# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payments": {
      "get": {
        "operationId": "PaymentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List payments for the store",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPaymentsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/pending-review": {
      "get": {
        "operationId": "PaymentsController_listPendingReview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List payments pending review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPaymentsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/order/{orderId}": {
      "get": {
        "operationId": "PaymentsController_getByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get payment by order ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}": {
      "get": {
        "operationId": "PaymentsController_getById",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get payment by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/submit-proof": {
      "post": {
        "operationId": "PaymentsController_submitProof",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/resubmit-proof": {
      "post": {
        "operationId": "PaymentsController_resubmitProof",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/start-review": {
      "post": {
        "operationId": "PaymentsController_startReview",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/approve": {
      "post": {
        "operationId": "PaymentsController_approve",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/reject": {
      "post": {
        "operationId": "PaymentsController_reject",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/collect-cod": {
      "post": {
        "operationId": "PaymentsController_collectCod",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/expire": {
      "post": {
        "operationId": "PaymentsController_expire",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{paymentId}/cancel": {
      "post": {
        "operationId": "PaymentsController_cancel",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "General Ecommerce API",
    "description": "General Ecommerce backend APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "AllowedTransitionDto": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "toStatus": {
            "type": "string"
          },
          "requiresReason": {
            "type": "boolean"
          }
        },
        "required": [
          "command",
          "toStatus",
          "requiresReason"
        ]
      },
...
      "PaymentWithOrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "submitted",
              "under_review",
              "approved",
              "rejected",
              "expired",
              "cancelled",
              "partially_refunded",
              "refunded"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "paidAmount": {
            "type": "string"
          },
          "refundedAmount": {
            "type": "string"
          },
          "refundableAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "submissionVersion": {
            "type": "number"
          },
          "storePaymentMethodId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCatalogId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCode": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodName": {
            "type": "string",
            "nullable": true
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "iban": {
            "type": "string",
            "nullable": true
          },
          "instructionsAr": {
            "type": "string",
            "nullable": true
          },
          "instructionsEn": {
            "type": "string",
            "nullable": true
          },
          "payerReference": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptUrl": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "payerNote": {
            "type": "string",
            "nullable": true
          },
          "customerSubmittedAt": {
            "type": "string",
            "nullable": true
          },
          "receiptUrl": {
            "type": "string",
            "nullable": true
          },
          "receiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "nullable": true
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          },
          "reviewNote": {
            "type": "string",
            "nullable": true
          },
          "customerUploadedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "allowedTransitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllowedTransitionDto"
            }
          },
          "orderNumber": {
            "type": "string"
          },
          "orderStatus": {
            "type": "string"
          },
          "orderTotal": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "method",
          "status",
          "statusLabel",
          "amount",
          "paidAmount",
          "refundedAmount",
          "refundableAmount",
          "currency",
          "version",
          "submissionVersion",
          "storePaymentMethodId",
          "paymentMethodCatalogId",
          "paymentMethodCode",
          "paymentMethodName",
          "accountName",
          "accountNumber",
          "phoneNumber",
          "iban",
          "instructionsAr",
          "instructionsEn",
          "payerReference",
          "payerReceiptUrl",
          "payerReceiptMediaAssetId",
          "payerNote",
          "customerSubmittedAt",
          "receiptUrl",
          "receiptMediaAssetId",
          "reviewedAt",
          "reviewedBy",
          "reviewNote",
          "customerUploadedAt",
          "createdAt",
          "updatedAt",
          "allowedTransitions",
          "orderNumber",
          "orderStatus",
          "orderTotal"
        ]
      },
      "PaginatedPaymentsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentWithOrderDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "submitted",
              "under_review",
              "approved",
              "rejected",
              "expired",
              "cancelled",
              "partially_refunded",
              "refunded"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "paidAmount": {
            "type": "string"
          },
          "refundedAmount": {
            "type": "string"
          },
          "refundableAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "submissionVersion": {
            "type": "number"
          },
          "storePaymentMethodId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCatalogId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCode": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodName": {
            "type": "string",
            "nullable": true
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "iban": {
            "type": "string",
            "nullable": true
          },
          "instructionsAr": {
            "type": "string",
            "nullable": true
          },
          "instructionsEn": {
            "type": "string",
            "nullable": true
          },
          "payerReference": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptUrl": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "payerNote": {
            "type": "string",
            "nullable": true
          },
          "customerSubmittedAt": {
            "type": "string",
            "nullable": true
          },
          "receiptUrl": {
            "type": "string",
            "nullable": true
          },
          "receiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "nullable": true
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          },
          "reviewNote": {
            "type": "string",
            "nullable": true
          },
          "customerUploadedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "allowedTransitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllowedTransitionDto"
            }
          }
        },
        "required": [
          "id",
          "orderId",
          "method",
          "status",
          "statusLabel",
          "amount",
          "paidAmount",
          "refundedAmount",
          "refundableAmount",
          "currency",
          "version",
          "submissionVersion",
          "storePaymentMethodId",
          "paymentMethodCatalogId",
          "paymentMethodCode",
          "paymentMethodName",
          "accountName",
          "accountNumber",
          "phoneNumber",
          "iban",
          "instructionsAr",
          "instructionsEn",
          "payerReference",
          "payerReceiptUrl",
          "payerReceiptMediaAssetId",
          "payerNote",
          "customerSubmittedAt",
          "receiptUrl",
          "receiptMediaAssetId",
          "reviewedAt",
          "reviewedBy",
          "reviewNote",
          "customerUploadedAt",
          "createdAt",
          "updatedAt",
          "allowedTransitions"
        ]
      },
      "PaymentCommandDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}